requirements.txt
alembic==1.13.1
aniso8601==9.0.1
Babel==2.14.0
bcrypt==4.1.3
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
Flask==3.0.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-MySQL==1.5.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
PyMySQL==1.1.0
pytest==8.2.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.29
stripe==9.9.0
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
WTForms==3.1.2
zipp==3.18.1
